+++ ICD.Connect.Settings/Core/ICoreSettings.cs
@@ -1,4 +1,6 @@
-﻿namespace ICD.Connect.Settings.Core
+﻿using ICD.Connect.Settings.Header;
+
+namespace ICD.Connect.Settings.Core
 {
 	public interface ICoreSettings : ISettings
 	{
@@ -6,5 +8,12 @@
 		/// Gets the child originator settings collection.
 		/// </summary>
 		SettingsCollection OriginatorSettings { get; }
+
+		/// <summary>
+		/// Parses and returns only the header portion from the full XML config.
+		/// </summary>
+		/// <param name="configXml"></param>
+		/// <returns></returns>
+		ConfigurationHeader GetHeader(string configXml);
 	}
 }